+++ rollup.config.js
@@ -1,3 +1,7 @@
+<<<<<<< HEAD
+=======
+import buble from '@rollup/plugin-buble'
+>>>>>>> upstream/4.0
 import replace from '@rollup/plugin-replace'
 import resolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
@@ -5,11 +9,18 @@ import { terser } from 'rollup-plugin-terser'
 import pkg from './package.json'
 
 const banner = `/*!
+<<<<<<< HEAD
  /**
   * vuex v${pkg.version}
   * (c) ${new Date().getFullYear()} Evan You
   * @license MIT
   */`
+=======
+ * vuex v${pkg.version}
+ * (c) ${new Date().getFullYear()} Evan You
+ * @license MIT
+ */`
+>>>>>>> upstream/4.0
 
 export function createEntries(configs) {
   return configs.map((c) => createEntry(c))
@@ -40,11 +51,23 @@ function createEntry(config) {
   }
 
   c.plugins.push(replace({
+<<<<<<< HEAD
     __DEV__: config.format === 'es' && !config.browser
+=======
+    __VERSION__: pkg.version,
+    __DEV__: config.format !== 'iife' && !config.browser
+>>>>>>> upstream/4.0
       ? `(process.env.NODE_ENV !== 'production')`
       : config.env !== 'production'
   }))
 
+<<<<<<< HEAD
+=======
+  if (config.transpile !== false) {
+    c.plugins.push(buble())
+  }
+
+>>>>>>> upstream/4.0
   c.plugins.push(resolve())
   c.plugins.push(commonjs())