+++ src/store.js
@@ -10,7 +10,7 @@ export function createStore (options) {
 
 export class Store {
   constructor (options = {}) {
-    if (process.env.NODE_ENV !== 'production') {
+    if (__DEV__) {
       assert(typeof Promise !== 'undefined', `vuex requires a Promise polyfill in this browser.`)
       assert(this instanceof Store, `store must be called with the new operator.`)
     }
@@ -148,6 +148,7 @@ export class Store {
       ? Promise.all(entry.map(handler => handler(payload)))
       : entry[0](payload)
 
+<<<<<<< HEAD
     return result.then(res => {
       try {
         this._actionSubscribers
@@ -157,9 +158,34 @@ export class Store {
         if (__DEV__) {
           console.warn(`[vuex] error in after action subscribers: `)
           console.error(e)
+=======
+    return new Promise((resolve, reject) => {
+      result.then(res => {
+        try {
+          this._actionSubscribers
+            .filter(sub => sub.after)
+            .forEach(sub => sub.after(action, this.state))
+        } catch (e) {
+          if (__DEV__) {
+            console.warn(`[vuex] error in after action subscribers: `)
+            console.error(e)
+          }
+>>>>>>> upstream/4.0
         }
-      }
-      return res
+        resolve(res)
+      }, error => {
+        try {
+          this._actionSubscribers
+            .filter(sub => sub.error)
+            .forEach(sub => sub.error(action, this.state, error))
+        } catch (e) {
+          if (__DEV__) {
+            console.warn(`[vuex] error in error action subscribers: `)
+            console.error(e)
+          }
+        }
+        reject(error)
+      })
     })
   }
 
@@ -217,7 +243,11 @@ export class Store {
   hasModule (path) {
     if (typeof path === 'string') path = [path]
 
+<<<<<<< HEAD
     if (process.env.NODE_ENV !== 'production') {
+=======
+    if (__DEV__) {
+>>>>>>> upstream/4.0
       assert(Array.isArray(path), `module path must be a string or an Array.`)
     }