+++ types/index.d.ts
@@ -42,6 +42,11 @@ export declare class Store<S> {
 }
 
 export function createStore<S>(options: StoreOptions<S>): Store<S>;
+<<<<<<< HEAD
+=======
+
+export function useStore<S = any>(): Store<S>;
+>>>>>>> upstream/4.0
 
 export interface Dispatch {
   (type: string, payload?: any, options?: DispatchOptions): Promise<any>;
@@ -79,10 +84,12 @@ export interface SubscribeOptions {
 }
 
 export type ActionSubscriber<P, S> = (action: P, state: S) => any;
+export type ActionErrorSubscriber<P, S> = (action: P, state: S, error: Error) => any;
 
 export interface ActionSubscribersObject<P, S> {
   before?: ActionSubscriber<P, S>;
   after?: ActionSubscriber<P, S>;
+  error?: ActionErrorSubscriber<P, S>;
 }
 
 export type SubscribeActionOptions<P, S> = ActionSubscriber<P, S> | ActionSubscribersObject<P, S>;