+++ types/vue.d.ts
@@ -2,7 +2,11 @@
  * Extends interfaces in Vue.js
  */
 
+<<<<<<< HEAD
 import { ComponentCustomOptions, ComponentCustomProperties } from "vue";
+=======
+import { ComponentCustomOptions } from "vue";
+>>>>>>> upstream/4.0
 import { Store } from "./index";
 
 declare module "@vue/runtime-core" {